{"schema":"decodex/commit/1","summary":"Split quantitative aggregate confidence","authority":"manual"}
